+++ plugin/src/main/kotlin/com/bytedance/android/plugin/AabResGuardPlugin.kt
@@ -9,6 +9,7 @@ import com.bytedance.android.plugin.tasks.AabResGuardTask
 import org.gradle.api.GradleException
 import org.gradle.api.Plugin
 import org.gradle.api.Project
+import java.io.File
 
 class AabResGuardPlugin : Plugin<Project> {
 
@@ -26,16 +27,26 @@ class AabResGuardPlugin : Plugin<Project> {
 
             val extension = project.extensions.getByType(AabResGuardExtension::class.java)
             val bundleFile = variant.artifacts.get(SingleArtifact.BUNDLE)
+            val outputDirectory = bundleFile.map { it.asFile.parentFile }
             val taskProvider = project.tasks.register(taskName, AabResGuardTask::class.java) { task ->
                 task.variantName.set(variant.name)
                 task.bundleFile.set(bundleFile)
                 task.obfuscatedBundleFile.set(
-                    project.layout.buildDirectory.file(
-                        "outputs/aabresguard/${variant.name}/${extension.obfuscatedBundleFileName}"
+                    project.layout.file(
+                        outputDirectory.map { directory ->
+                            File(directory, extension.obfuscatedBundleFileName)
+                        }
+                    )
+                )
+                task.resourceMappingFile.set(
+                    project.layout.file(
+                        outputDirectory.map { directory ->
+                            File(directory, "resources-mapping.txt")
+                        }
                     )
                 )
                 task.enableObfuscate.set(extension.enableObfuscate)
-                task.mappingFile.fileValue(extension.mappingFile?.toFile())
+                task.mappingFilePath.set(extension.mappingFile?.toString().orEmpty())
                 task.whiteList.set(extension.whiteList.orEmpty())
                 task.mergeDuplicatedRes.set(extension.mergeDuplicatedRes)
                 task.enableFilterFiles.set(extension.enableFilterFiles)